requirements.txt
# requirements.txt
aiofiles==23.2.1
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
beautifulsoup4==4.12.2
black==23.11.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
dlib==19.22.0
exceptiongroup==1.2.0
fastapi==0.104.1
ffmpy==0.3.1
filelock==3.13.1
fonttools==4.45.0
fsspec==2023.10.0
ftfy==6.1.3
gdown==4.7.1
gradio==4.5.0
gradio-client==0.7.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
huggingface-hub==0.19.4
idna==3.4
importlib-resources==6.1.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.4
mdurl==0.1.2
mypy-extensions==1.0.0
numpy==1.24.4
openai-clip==1.0.1
orjson==3.9.10
packaging==23.2
pandas==2.0.3
pathspec==0.11.2
Pillow==10.1.0
pkgutil-resolve-name==1.3.10
platformdirs==4.0.0
plotly==5.18.0
pydantic==2.5.2
pydantic-core==2.14.5
pydub==0.25.1
pygments==2.17.2
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.0
regex==2023.10.3
requests==2.31.0
rich==13.7.0
rpds-py==0.13.1
scikit-learn==1.3.2
scipy==1.10.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.27.0
tenacity==8.2.3
threadpoolctl==3.2.0
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.0
torchdiffeq==0.2.3
tqdm==4.66.1
typer==0.9.0
typing-extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.24.0.post1
wcwidth==0.2.12
websockets==11.0.3
zipp==3.17.0
seaborn
adversarial-robustness-toolbox
facenet-pytorch
pytorch_msssim
